IRIS-CLASSIFICATION
Visualizations including scatter plots and bar plots were generated to gain insights into the distribution of the features and the species of Iris flowers present in the dataset. Utilizing the K-nearest neighbors classifier, I trained the model on the training data and made predictions on the testing data. calculated the accuracy score. visualized the confusion matrix to gain a deeper understanding of the model's classification results.